backend/requirements.txt
fastapi
uvicorn
websockets
pydantic
python-dotenv
python-multipart
opencv-python
mediapipe
firebase-admin
httpx
numpy
groq
gTTS

